requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.8.1=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
blas=1.0=mkl
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
bzip2=1.0.8=h6c40b1e_5
ca-certificates=2023.12.12=hecd8cb5_0
contourpy=1.2.0=py311ha357a0b_0
cycler=0.11.0=pyhd3eb1b0_0
decorator=5.1.1=pyhd3eb1b0_0
django=5.0.4=pypi_0
django-cors-headers=4.3.1=pypi_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
imageio=2.33.1=py311hecd8cb5_0
intel-openmp=2023.1.0=ha357a0b_43548
ipython=8.20.0=py311hecd8cb5_0
jedi=0.18.1=py311hecd8cb5_1
jpeg=9e=h6c40b1e_1
kiwisolver=1.4.4=py311hcec6c5f_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libffi=3.4.4=hecd8cb5_0
libpng=1.6.39=h6c40b1e_0
libtiff=4.5.1=hcec6c5f_0
libwebp-base=1.3.2=h6c40b1e_0
lz4-c=1.9.4=hcec6c5f_0
matplotlib=3.8.0=py311hecd8cb5_0
matplotlib-base=3.8.0=py311h41a4f6b_0
matplotlib-inline=0.1.6=py311hecd8cb5_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py311h6c40b1e_1
mkl_fft=1.3.8=py311h6c40b1e_0
mkl_random=1.2.4=py311ha357a0b_0
munkres=1.1.4=py_0
ncurses=6.4=hcec6c5f_0
numpy=1.26.4=py311h728a8a3_0
numpy-base=1.26.4=py311h53bf9ac_0
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1w=hca72f7f_0
packaging=23.1=py311hecd8cb5_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.2.0=py311h6c40b1e_0
pip=23.3.1=py311hecd8cb5_0
prompt-toolkit=3.0.43=py311hecd8cb5_0
prompt_toolkit=3.0.43=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pygments=2.15.1=py311hecd8cb5_1
pyparsing=3.0.9=py311hecd8cb5_0
python=3.11.2=h1fd4e5f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
readline=8.2=hca72f7f_0
setuptools=68.2.2=py311hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h6c40b1e_0
sqlparse=0.5.0=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.8.0=ha357a0b_0
tk=8.6.12=h5d9f67b_0
tornado=6.3.3=py311h6c40b1e_0
traitlets=5.7.1=py311hecd8cb5_0
tzdata=2024a=h04d1e81_0
uuid=1.30=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.41.2=py311hecd8cb5_0
xz=5.4.6=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
